moved item editing from item store to single item store
