+++ client/components/EditItem.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
-import {modItem} from '../store/item'
+import {modItem} from '../store/singleItem'
 
 class EditItem extends Component {
   constructor() {
